fix(ui): 당첨 통계 메서드의 메개변수의 지네릭스 value의 타입을 Long으로 변경

+++ src/main/java/lotto/ui/OutputPrinter.java
@@ -17,7 +17,7 @@ public class OutputPrinter {
         }
     }
 
-    public void printWinningStatistics(Map<WinningNumbers, Integer> winningCase) {
+    public void printWinningStatistics(Map<WinningNumbers, Long> winningCase) {
         System.out.println("당첨 통계");
         System.out.println("---");
